Create Let's play Guess the Number!
In this code, the random.randint() function is used to generate a random number between 1 and 100, which is stored in the number variable. The while loop continues until the player correctly guesses the number, which sets game_over to True.

Inside the loop, the player is prompted to enter a guess using the input() function, and the guess is converted to an integer using the int() function. The guess is then compared to the random number using an if statement, and the computer prints a message telling the player whether their guess was too high, too low, or correct. If the guess is correct, the loop ends and the player wins the game.